Add warning options, GPLv3 license headers

+++ src/discover.c
@@ -1,3 +1,17 @@
+/**
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 
 #include <stdio.h>
 #include <netinet/in.h>
@@ -54,7 +68,7 @@ void ps4_discover(const char *host_name)
 
 	char buffer[512];
 	int len = ps4_discovery_write_packet(&packet, buffer, sizeof(buffer));
-	if(len > sizeof(buffer))
+	if(len > (int)sizeof(buffer))
 		return;
 	ssize_t rc = sendto(s, buffer, (size_t)len+1, 0, (struct sockaddr *)&server_addr, sizeof(server_addr));
